+++ app/src/main/java/com/example/gagyeboost/model/data/DateItem.kt
@@ -20,7 +20,7 @@ data class DateItem(
 )
 
 data class Filter(
-    val moneyType: Int?,
+    val moneyType: Byte?,
     val startYear: Int,
     val startMonth: Int,
     val startDay: Int,
